+++ nes-runtime/include/Nautilus/IR/BasicBlocks/BasicBlock.hpp
@@ -37,6 +37,21 @@ class BasicBlock : public std::enable_shared_from_this<BasicBlock> {
     [[nodiscard]] std::string getIdentifier();
     [[nodiscard]] uint32_t getScopeLevel();
     void setScopeLevel(uint32_t scopeLevel);
+
+    /**
+     * @brief Get the number of edges that lead back from the loop body to the loop header.
+     */
+    [[nodiscard]] uint32_t getNumLoopBackEdges();
+
+    /**
+     * @brief Increment counter for edges that lead back from the loop body to the loop header.
+     */
+    void incrementNumLoopBackEdge();
+
+    /**
+     * @brief Check if the counter for edges that lead back from the loop body to the loop header is > 0.
+     */
+    [[nodiscard]] bool isLoopHeaderBlock();
     [[nodiscard]] std::vector<Operations::OperationPtr> getOperations();
     [[nodiscard]] Operations::OperationPtr getTerminatorOp();
     [[nodiscard]] std::vector<std::shared_ptr<Operations::BasicBlockArgument>> getArguments();
@@ -58,6 +73,7 @@ class BasicBlock : public std::enable_shared_from_this<BasicBlock> {
   private:
     std::string identifier;
     uint32_t scopeLevel;
+    uint32_t numLoopBackEdges;
     std::vector<Operations::OperationPtr> operations;
     std::vector<std::shared_ptr<Operations::BasicBlockArgument>> arguments;
     std::vector<std::weak_ptr<BasicBlock>> predecessors;