+++ nes-runtime/include/Nautilus/Interface/DataTypes/Text/Text.hpp
@@ -43,6 +43,20 @@ class Text final : public Nautilus::Any {
     Value<Boolean> equals(const Value<Text>& other) const;
 
     /**
+     * @brief concat two text object
+     * @param other text object
+     * @return combination of two text object
+     */
+    Value<Text> concat(const Value<Text>& other) const;
+
+    /**
+     * @brief  Extract a subText of length len from the startIndex.Note that the start value of 1 refers to the first character of the Text
+     * @param  Value startIndex, Value len
+     * @return a subtext object
+     */
+    Value<Text> substring(Value<UInt32>, Value<UInt32>) const;
+
+    /**
      * @brief Returns the number of characters of this text value.
      * @return Value<Int32>
      */
@@ -93,5 +107,8 @@ requires std::is_same_v<TextValue*, T> auto createDefault() {
     auto text = Value<Text>(std::make_unique<Text>(textRef));
     return text;
 }
+
+Value<Text> transformReturnValues(TextValue* value);
+
 }// namespace NES::Nautilus
 #endif// NES_RUNTIME_INCLUDE_NAUTILUS_INTERFACE_DATATYPES_TEXT_TEXT_HPP_