+++ nes-runtime/include/Runtime/TupleBuffer.hpp
@@ -85,6 +85,14 @@ class TupleBuffer {
     [[nodiscard]] constexpr TupleBuffer() noexcept = default;
 
     /**
+     * @brief Interprets the void* as a pointer to the content of tuple buffer
+     * @note if bufferPointer is not pointing to the begin of an data buffer the behavior of this function is undefined.
+     * @param bufferPointer
+     * @return TupleBuffer
+     */
+    [[maybe_unused]] static TupleBuffer reinterpretAsTupleBuffer(void* bufferPointer);
+
+    /**
      * @brief Creates a TupleBuffer of length bytes starting at ptr address.
      *
      * @param ptr    resource's address.